pyproject.toml
[tool.poetry]
name = "rttb-simple"
version = "0.1.0"
description = ""
authors = ["Jan Marais <maraisjandre9@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
httpx = "^0.21.1"
python-dotenv = "^0.19.2"
Jinja2 = "^3.0.3"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
black = "^21.11b1"
ipykernel = "^6.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
